Add JS API spec text
In the overview, the `Memory.buffer` getter only creates a new
`SharedArrayBuffer` when the lengths don't match. In the JS spec, there
is a Memory object cache which is unconditionally updated with the most
recent memory object and its length after every grow operation. So I
think it makes sense to specify that the `buffer` getter always accesses
this cache, since it is simpler (and unobservable, I think).